Introduce permission and permission_action models
Previously we had an action function as an attribute of a role permission, which made role configuration (a user activity, albeit a sysadmin one) extremely difficult. Now instead we can predefine more intuitive permissions, each of which is associated behind the scenes with a set of action functions. So establishing the relationships between actions and permissions becomes a developer rather than a sysadmin activity, which is correct since actions are code objects.