Do not check client_id when validating access tokens
It is not necessary, and is actually undesirable in terms of OAuth2, which requires only that a token have a valid scope (and audience, expiry, etc) for the API to which it requires access. A valid scope implies that the client is permitted to access the API. The API does not, and should not, have to know in advance specifically which clients will be calling it.